refactor: metrics::metrics is now metrics
